Move bank and basket api queries to it's own file
