+++ web-registry/src/lib/ecocredit.ts
@@ -2,18 +2,6 @@ import axios, { AxiosResponse } from 'axios';
 import { TxResponse } from '@regen-network/api/lib/generated/cosmos/base/abci/v1beta1/abci';
 import qs from 'querystring';
 
-import {
-  QueryClientImpl as BankQueryClient,
-  QueryDenomMetadataResponse,
-  QueryBalanceResponse,
-} from '@regen-network/api/lib/generated/cosmos/bank/v1beta1/query';
-import {
-  QueryClientImpl as BasketQueryClient,
-  QueryBasketResponse,
-  QueryBasketsResponse,
-  QueryBasketBalancesResponse,
-} from '@regen-network/api/lib/generated/regen/ecocredit/basket/v1/query';
-
 import { expLedger, ledgerRESTUri } from '../lib/ledger';
 import type { PageResponse } from '../types/ledger/base';
 import type {
@@ -356,135 +344,5 @@ export const queryEcoClassInfo = async (
     return data;
   } catch (err) {
     throw new Error(`Error fetching class info: ${err}`);
-  }
-};
-
-/**
- *
- * REGEN LEDGER API QUERIES
- *
- */
-
-/**
- *
- * BANK MODULE QUERIES
- * -------------------
- *  - AllBalances (TODO)
- *  - Balance
- *  - DenomMetadata
- *  - DenomsMetadata (TODO)
- *
- */
-
-// Denom Metadata
-
-export type QueryDenomMetadataProps = {
-  client: BankQueryClient;
-  denom: string;
-};
-
-export const queryDenomMetadata = async ({
-  client,
-  denom,
-}: QueryDenomMetadataProps): Promise<
-  QueryDenomMetadataResponse | undefined
-> => {
-  try {
-    const metadata = await client.DenomMetadata({ denom });
-    return metadata;
-  } catch (err) {
-    // return err as Error;
-    // throw new Error('queryDenomMetadata');
-    return;
-  }
-};
-
-// Balance
-
-export type QueryBalanceProps = {
-  client: BankQueryClient;
-  address: string;
-  denom: string;
-};
-
-export const queryBalance = async ({
-  client,
-  address,
-  denom,
-}: QueryBalanceProps): Promise<QueryBalanceResponse | undefined> => {
-  try {
-    const balance = await client.Balance({ address, denom });
-    return balance;
-  } catch (err) {
-    // return err as Error;
-    // throw new Error('queryBalance');
-    return;
-  }
-};
-
-/**
- *
- * BASKET MODULE QUERIES
- * ---------------------
- *  - Basket
- *  - Baskets
- *  - BasketBalances
- *  - BasketBalance (TODO)
- *
- */
-
-// Basket
-
-export type QueryBasketProps = {
-  client: BasketQueryClient;
-  basketDenom: string;
-  // params: {
-  //   basketDenom: string;
-  // };
-};
-
-export const queryBasket = async ({
-  client,
-  basketDenom,
-}: // params: { basketDenom },
-QueryBasketProps): Promise<QueryBasketResponse | Error> => {
-  try {
-    return await client.Basket({ basketDenom });
-  } catch (err) {
-    return err as Error;
-  }
-};
-
-// Baskets
-
-export type QueryBasketsProps = {
-  client: BasketQueryClient;
-};
-
-export const queryBaskets = async ({
-  client,
-}: QueryBasketsProps): Promise<QueryBasketsResponse | Error> => {
-  try {
-    return await client.Baskets({});
-  } catch (err) {
-    return err as Error;
-  }
-};
-
-// Basket Balances
-
-export type QueryBasketBalancesProps = {
-  client: BasketQueryClient;
-  basketDenom: string;
-};
-
-export const queryBasketBalances = async ({
-  client,
-  basketDenom,
-}: QueryBasketBalancesProps): Promise<QueryBasketBalancesResponse | Error> => {
-  try {
-    return await client.BasketBalances({ basketDenom });
-  } catch (err) {
-    return err as Error;
   }
 };
